move macro expanded code into comment
